Windows port: experimental ccl: MSVC narrowing casts and pimpl-completeness fixes
- all_gather_async, all_gather_concat_heads_fused, all_gather_minimal_matmul_async,
  all_reduce_async, all_to_all_async(_generic), all_to_all_dispatch_metadata,
  deepseek_moe_reduce_scatter, llama_all_gather_matmul_async, llama_reduce_scatter(_create_heads),
  llama_reduce_scatter_matmul, minimal_matmul_strided_reduce_scatter_async, moe_compute,
  neighbor_pad_async, neighbor_pad_halo, reduce_scatter_minimal_async(_direct),
  ring_attention_all_gather_async, send_recv_async (recv/send, direct and indirect),
  slice_reshard_async, strided_all_gather_async, strided_reduce_scatter_async:
  add static_cast<uint32_t>(...) around DeviceAddr/size_t/int expressions
  (.buffer()->address(), .address(), .size(), CoreCoord.x/.y, etc.) that MSVC treats
  as narrowing inside std::vector<uint32_t> and other braced-init-lists used for
  compile-time and runtime kernel args.
- ring_attention_all_gather_async: fix two CBDescriptor aggregate-init failures
  (C2440) caused by size_t packet-header-size expressions narrowing into
  uint32_t total_size/page_size fields.
- all_gather_async, all_to_all_dispatch_metadata, llama_reduce_scatter(_create_heads),
  minimal_matmul_strided_reduce_scatter_async, moe/selective_reduce_combine,
  reduce_scatter_minimal_async, ring_attention_all_gather_async, strided_all_gather_async,
  strided_reduce_scatter_async, dit_fused_distributed_groupnorm/rmsnorm,
  all_gather_minimal_matmul_async: include tt_metal/impl/buffers/global_semaphore_impl.hpp
  wherever operation_attributes_t (or a struct it nests) stores a GlobalSemaphore by
  value (in std::optional or std::vector). MSVC's aggregate/attribute reflection
  machinery in device_operation.hpp and tt_stl/reflection.hpp needs the
  GlobalSemaphoreImpl pimpl type to be complete at these points; GCC/Clang do not.

+++ ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_async/device/all_gather_async_default_program_factory.cpp
@@ -707,9 +707,9 @@ AllGatherProgramArtifacts build_all_gather_async_minimal_default_program_artifac
                 }
 
                 std::vector<uint32_t> reader_rt_args = {
-                    input_tensor.buffer()->address(),   // input_tensor_address
-                    output_tensor.buffer()->address(),  // output_tensor_address
-                    semaphore.at(dir).address(),        // out_ready_sem
+                    static_cast<uint32_t>(input_tensor.buffer()->address()),   // input_tensor_address
+                    static_cast<uint32_t>(output_tensor.buffer()->address()),  // output_tensor_address
+                    static_cast<uint32_t>(semaphore.at(dir).address()),        // out_ready_sem
                     dir,                                // direction RT ARG
                     input_tile_id_start,                // input_tile_id_start RT ARG
                     input_tile_id_end,                  // input_tile_id_end RT ARG
@@ -745,16 +745,16 @@ AllGatherProgramArtifacts build_all_gather_async_minimal_default_program_artifac
                     mesh_device->worker_core_from_logical_core(termination_master_logical_core);
 
                 std::vector<uint32_t> writer_rt_args = {
-                    output_tensor.buffer()->address(),                           // output_tensor_address
-                    virtual_core.x,                                              // out_ready_sem_noc0_x
-                    virtual_core.y,                                              // out_ready_sem_noc0_y
-                    semaphore.at(dir).address(),                                 // out_ready_sem
+                    static_cast<uint32_t>(output_tensor.buffer()->address()),    // output_tensor_address
+                    static_cast<uint32_t>(virtual_core.x),                       // out_ready_sem_noc0_x
+                    static_cast<uint32_t>(virtual_core.y),                       // out_ready_sem_noc0_y
+                    static_cast<uint32_t>(semaphore.at(dir).address()),          // out_ready_sem
                     barrier_semaphore.has_value() && !using_persistent_buffers,  // use synchronize barrier semaphore
                     barrier_semaphore.has_value()                                // synchronize barrier semaphore
-                        ? barrier_semaphore.value().address()
+                        ? static_cast<uint32_t>(barrier_semaphore.value().address())
                         : 0,
-                    opposite_core_coord.x,    // opposite_core_sem_noc0_x
-                    opposite_core_coord.y,    // opposite_core_sem_noc0_y
+                    static_cast<uint32_t>(opposite_core_coord.x),  // opposite_core_sem_noc0_x
+                    static_cast<uint32_t>(opposite_core_coord.y),  // opposite_core_sem_noc0_y
                     dir,                      // direction
                     input_tile_id_start,      // input_tile_id_start
                     input_tile_id_end,        // input_tile_id_end
